Make string handling unicode-preserving
Previously, it was impossible to output non-ASCII characters using cecho for
multiple reasons:

 * Strings were processed as flattened lists.
   This caused encoding information to not be preserved properly.
 * Data was transmitted as string from the Erlang to the C side causing
   encoding information not to arrive correctly.
 * The library linked against libncurses rather than libncursesw and as
   a result, no “wide character” support was available.

This commit fixes this by switching the string handling to be based on iolists
and binaries. String values are transferred to the C-side as binaries rather
than strings now.

Additionally, upon compilation, the C part is linked against `libncursesw` in
favor of the previously chosen `libncurses`.

The API remains compatible with preceding invocations and still allows strings
to be passed to all of the string functions.